Move TenantId and UserId to SharedKernel
