+++ application/account-management/Api/Endpoints/UserEndpoints.cs
@@ -1,8 +1,8 @@
 using PlatformPlatform.AccountManagement.Users.Commands;
-using PlatformPlatform.AccountManagement.Users.Domain;
 using PlatformPlatform.AccountManagement.Users.Queries;
 using PlatformPlatform.SharedKernel.ApiResults;
 using PlatformPlatform.SharedKernel.Endpoints;
+using PlatformPlatform.SharedKernel.Entities;
 
 namespace PlatformPlatform.AccountManagement.Api.Endpoints;
