+++ application/account-management/Core/Authentication/Commands/RefreshAuthenticationTokens.cs
@@ -6,6 +6,7 @@ using PlatformPlatform.AccountManagement.Authentication.Services;
 using PlatformPlatform.AccountManagement.TelemetryEvents;
 using PlatformPlatform.AccountManagement.Users.Domain;
 using PlatformPlatform.SharedKernel.Cqrs;
+using PlatformPlatform.SharedKernel.Entities;
 using PlatformPlatform.SharedKernel.TelemetryEvents;
 
 namespace PlatformPlatform.AccountManagement.Authentication.Commands;
