+++ application/account-management/Core/Signups/Domain/Signup.cs
@@ -1,5 +1,4 @@
 using JetBrains.Annotations;
-using PlatformPlatform.AccountManagement.Tenants.Domain;
 using PlatformPlatform.SharedKernel.Entities;
 using PlatformPlatform.SharedKernel.IdGenerators;
