+++ application/account-management/Core/Signups/Domain/SignupRepository.cs
@@ -1,5 +1,4 @@
 using PlatformPlatform.AccountManagement.Database;
-using PlatformPlatform.AccountManagement.Tenants.Domain;
 using PlatformPlatform.SharedKernel.Entities;
 using PlatformPlatform.SharedKernel.Persistence;
