+++ application/account-management/Core/TelemetryEvents/TelemetryEvents.cs
@@ -1,5 +1,6 @@
 using PlatformPlatform.AccountManagement.Tenants.Domain;
 using PlatformPlatform.AccountManagement.Users.Domain;
+using PlatformPlatform.SharedKernel.Entities;
 using PlatformPlatform.SharedKernel.TelemetryEvents;
 
 namespace PlatformPlatform.AccountManagement.TelemetryEvents;
