+++ application/account-management/Core/Users/Commands/RemoveAvatar.cs
@@ -2,6 +2,7 @@ using JetBrains.Annotations;
 using PlatformPlatform.AccountManagement.TelemetryEvents;
 using PlatformPlatform.AccountManagement.Users.Domain;
 using PlatformPlatform.SharedKernel.Cqrs;
+using PlatformPlatform.SharedKernel.Entities;
 using PlatformPlatform.SharedKernel.TelemetryEvents;
 
 namespace PlatformPlatform.AccountManagement.Users.Commands;
