+++ application/account-management/Core/Users/Domain/User.cs
@@ -1,4 +1,3 @@
-using PlatformPlatform.AccountManagement.Tenants.Domain;
 using PlatformPlatform.SharedKernel.Entities;
 
 namespace PlatformPlatform.AccountManagement.Users.Domain;
