+++ application/account-management/Core/Users/Domain/UserTypes.cs
@@ -1,18 +1,6 @@
 using JetBrains.Annotations;
-using PlatformPlatform.SharedKernel.IdGenerators;
 
 namespace PlatformPlatform.AccountManagement.Users.Domain;
-
-[PublicAPI]
-[IdPrefix("usr")]
-[TypeConverter(typeof(StronglyTypedIdTypeConverter<string, UserId>))]
-public sealed record UserId(string Value) : StronglyTypedUlid<UserId>(Value)
-{
-    public override string ToString()
-    {
-        return Value;
-    }
-}
 
 [PublicAPI]
 [JsonConverter(typeof(JsonStringEnumConverter))]
