+++ application/shared-kernel/SharedKernel/Entities/TenantId.cs
@@ -1,0 +1,38 @@
+using System.Diagnostics.CodeAnalysis;
+using PlatformPlatform.SharedKernel.IdGenerators;
+
+namespace PlatformPlatform.SharedKernel.Entities;
+
+[PublicAPI]
+[TypeConverter(typeof(StronglyTypedIdTypeConverter<string, TenantId>))]
+public sealed record TenantId(string Value) : StronglyTypedId<string, TenantId>(Value)
+{
+    public override string ToString()
+    {
+        return Value;
+    }
+
+    public static bool TryParse(string? value, [NotNullWhen(true)] out TenantId? result)
+    {
+        if (value is not { Length: >= 3 and <= 30 })
+        {
+            result = null;
+            return false;
+        }
+
+        if (!value.All(c => char.IsLower(c) || char.IsDigit(c) || char.IsDigit('-')))
+        {
+            result = null;
+            return false;
+        }
+
+        if (value.StartsWith('-') || value.EndsWith('-'))
+        {
+            result = null;
+            return false;
+        }
+
+        result = new TenantId(value);
+        return true;
+    }
+}
